-
trying to figure out animation. deleted a1 bc it had an issue. made 2 more scenes for testing while fixing animation. it doesnt save each time when i use the animator to make the transitions, need to figure out how to get it to save. changed the setting in animator for the sprite to keep animator state, but i think i'm missing something cuz it doesnt work.