worked on camera
set up cinemachine virtual camera with confiner for the worldboundary (which is a polygon collider asset on its own) and script for when player hits boundary it is sent to 0,0. (if time allots, could add larger world and make it so it naturally slows/stops player movement)

also placed 3 enemies around the world and set the npc in a location.